+++ src/main/java/SMU/BAMBOO/Hompage/global/upload/service/AwsS3Uploader.java
@@ -2,6 +2,7 @@ package SMU.BAMBOO.Hompage.global.upload.service;
 
 import SMU.BAMBOO.Hompage.global.exception.CustomException;
 import SMU.BAMBOO.Hompage.global.exception.ErrorCode;
+import SMU.BAMBOO.Hompage.global.upload.util.AwsS3FileNameGenerator;
 import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.model.CannedAccessControlList;
 import com.amazonaws.services.s3.model.ObjectMetadata;
