+++ src/main/java/SMU/BAMBOO/Hompage/global/upload/util/AwsS3FileValidator.java
@@ -1,11 +1,11 @@
-package SMU.BAMBOO.Hompage.global.upload.service;
+package SMU.BAMBOO.Hompage.global.upload.util;
 
 import SMU.BAMBOO.Hompage.global.exception.CustomException;
 import SMU.BAMBOO.Hompage.global.exception.ErrorCode;
-import org.springframework.stereotype.Service;
+import org.springframework.stereotype.Component;
 import org.springframework.web.multipart.MultipartFile;
 
-@Service
+@Component
 public class AwsS3FileValidator {
 
     public void validateFile(MultipartFile file) {
